move katana
