+++ account_sdk/src/providers/katana_runner.rs
@@ -1,0 +1,116 @@
+use serde::Deserialize;
+
+use std::path::{Path, PathBuf};
+use std::process::{Child, Command, Stdio};
+use std::sync::mpsc;
+use std::thread;
+use std::time::Duration;
+use std::{
+    fs,
+    io::{BufRead, BufReader, Write},
+};
+use std::{fs::File, process::ChildStdout, sync::mpsc::Sender};
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct KatanaRunnerConfig {
+    pub port: u16,
+    pub katana_path: String,
+    pub log_file_path: String,
+}
+
+impl KatanaRunnerConfig {
+    pub fn from_file(path: impl Into<String>) -> Self {
+        let config_string = fs::read_to_string(path.into()).expect("Failed to read config file");
+
+        toml::from_str(&config_string).expect("Failed to parse config file")
+    }
+    pub fn port(mut self, port: u16) -> Self {
+        self.port = port;
+        self.log_file_path = KatanaRunnerConfig::add_port_to_filename(&self.log_file_path, port);
+        self
+    }
+    fn add_port_to_filename(file_path: &str, port: u16) -> String {
+        let path = Path::new(file_path);
+        let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
+        let extension = path.extension().and_then(|s| s.to_str()).unwrap_or("");
+
+        let new_file_name = if extension.is_empty() {
+            format!("{}_{}", stem, port)
+        } else {
+            format!("{}_{}.{}", stem, port, extension)
+        };
+        let mut new_path = PathBuf::from(path.parent().unwrap_or_else(|| Path::new("")));
+        new_path.push(new_file_name);
+        new_path.to_string_lossy().into_owned()
+    }
+}
+
+#[derive(Debug)]
+pub struct KatanaRunner {
+    child: Child,
+    port: u16,
+}
+
+impl KatanaRunner {
+    pub fn new(config: KatanaRunnerConfig) -> Self {
+        let mut child = Command::new(config.katana_path)
+            .args(["-p", &config.port.to_string()])
+            .args(["--json-log"])
+            .stdout(Stdio::piped())
+            .spawn()
+            .expect("failed to start subprocess");
+
+        let stdout = child
+            .stdout
+            .take()
+            .expect("failed to take subprocess stdout");
+
+        let (sender, receiver) = mpsc::channel();
+
+        thread::spawn(move || {
+            KatanaRunner::wait_for_server_started_and_signal(&config.log_file_path, stdout, sender);
+        });
+
+        receiver
+            .recv_timeout(Duration::from_secs(5))
+            .expect("timeout waiting for server to start");
+
+        KatanaRunner {
+            child,
+            port: config.port,
+        }
+    }
+
+    fn wait_for_server_started_and_signal(
+        log_file_path: &str,
+        stdout: ChildStdout,
+        sender: Sender<()>,
+    ) {
+        let reader = BufReader::new(stdout);
+        let mut log_writer = File::create(log_file_path).expect("failed to create log file");
+
+        for line in reader.lines() {
+            let line = line.expect("failed to read line from subprocess stdout");
+            writeln!(log_writer, "{}", line).expect("failed to write to log file");
+
+            if line.contains(r#""target":"katana""#) {
+                sender.send(()).expect("failed to send start signal");
+            }
+        }
+    }
+
+    pub fn port(&self) -> u16 {
+        self.port
+    }
+}
+
+impl Drop for KatanaRunner {
+    fn drop(&mut self) {
+        if let Err(e) = self.child.kill() {
+            eprintln!("Failed to kill katana subprocess: {}", e);
+        }
+        if let Err(e) = self.child.wait() {
+            eprintln!("Failed to wait for katana subprocess: {}", e);
+        }
+    }
+}
