+++ account_sdk/src/providers/mod.rs
@@ -23,6 +23,8 @@ use crate::rpc_provider::RpcClientProvider;
 use lazy_static::lazy_static;
 
 mod devnet;
+mod katana;
+mod katana_runner;
 
 lazy_static! {
     pub static ref UDC_ADDRESS: FieldElement =
@@ -46,8 +48,6 @@ lazy_static! {
         .unwrap()
     );
 }
-
-pub use devnet::StarknetDevnet;
 
 pub trait PredeployedProvider
 where
@@ -73,141 +73,4 @@ impl PredeployedAccount {
 pub struct PredeployedContract {
     pub address: FieldElement,
     pub class_hash: FieldElement,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct KatanaRunnerConfig {
-    pub port: u16,
-    pub katana_path: String,
-    pub log_file_path: String,
-}
-
-impl KatanaRunnerConfig {
-    pub fn from_file(path: impl Into<String>) -> Self {
-        let config_string = fs::read_to_string(path.into()).expect("Failed to read config file");
-
-        toml::from_str(&config_string).expect("Failed to parse config file")
-    }
-    pub fn port(mut self, port: u16) -> Self {
-        self.port = port;
-        self.log_file_path = KatanaRunnerConfig::add_port_to_filename(&self.log_file_path, port);
-        self
-    }
-    fn add_port_to_filename(file_path: &str, port: u16) -> String {
-        let path = Path::new(file_path);
-        let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
-        let extension = path.extension().and_then(|s| s.to_str()).unwrap_or("");
-
-        let new_file_name = if extension.is_empty() {
-            format!("{}_{}", stem, port)
-        } else {
-            format!("{}_{}.{}", stem, port, extension)
-        };
-        let mut new_path = PathBuf::from(path.parent().unwrap_or_else(|| Path::new("")));
-        new_path.push(new_file_name);
-        new_path.to_string_lossy().into_owned()
-    }
-}
-
-#[derive(Debug)]
-pub struct KatanaRunner {
-    child: Child,
-    port: u16,
-}
-
-impl KatanaRunner {
-    pub fn new(config: KatanaRunnerConfig) -> Self {
-        let mut child = Command::new(config.katana_path)
-            .args(["-p", &config.port.to_string()])
-            .args(["--json-log"])
-            .stdout(Stdio::piped())
-            .spawn()
-            .expect("failed to start subprocess");
-
-        let stdout = child
-            .stdout
-            .take()
-            .expect("failed to take subprocess stdout");
-
-        let (sender, receiver) = mpsc::channel();
-
-        thread::spawn(move || {
-            KatanaRunner::wait_for_server_started_and_signal(&config.log_file_path, stdout, sender);
-        });
-
-        receiver
-            .recv_timeout(Duration::from_secs(5))
-            .expect("timeout waiting for server to start");
-
-        KatanaRunner {
-            child,
-            port: config.port,
-        }
-    }
-
-    fn wait_for_server_started_and_signal(
-        log_file_path: &str,
-        stdout: ChildStdout,
-        sender: Sender<()>,
-    ) {
-        let reader = BufReader::new(stdout);
-        let mut log_writer = File::create(log_file_path).expect("failed to create log file");
-
-        for line in reader.lines() {
-            let line = line.expect("failed to read line from subprocess stdout");
-            writeln!(log_writer, "{}", line).expect("failed to write to log file");
-
-            if line.contains(r#""target":"katana""#) {
-                sender.send(()).expect("failed to send start signal");
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct KatanaClientProvider {
-    port: u16,
-}
-
-// todo!("Implement starknet::providers::Provider");
-// impl Provider for KatanaClientProvider {}
-
-impl From<u16> for KatanaClientProvider {
-    fn from(value: u16) -> Self {
-        KatanaClientProvider { port: value }
-    }
-}
-
-impl From<&KatanaRunner> for KatanaClientProvider {
-    fn from(value: &KatanaRunner) -> Self {
-        KatanaClientProvider { port: value.port }
-    }
-}
-
-impl RpcClientProvider<HttpTransport> for KatanaClientProvider {
-    fn get_client(&self) -> JsonRpcClient<HttpTransport> {
-        JsonRpcClient::new(HttpTransport::new(
-            Url::parse(&format!("http://0.0.0.0:{}/", self.port)).unwrap(),
-        ))
-    }
-
-    fn chain_id(&self) -> FieldElement {
-        FieldElement::from_byte_slice_be(&"KATANA".as_bytes()[..]).unwrap()
-    }
-}
-
-impl Drop for KatanaRunner {
-    fn drop(&mut self) {
-        if let Err(e) = self.child.kill() {
-            eprintln!("Failed to kill katana subprocess: {}", e);
-        }
-        if let Err(e) = self.child.wait() {
-            eprintln!("Failed to wait for katana subprocess: {}", e);
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct StarknetDevnet {
-    pub port: u16,
 }
